Update main.py
Sooo I was having scope issues with getting the data from queries and I tried making connection and curs global via defining them at the top of the main file or using 'global connection' later on... and I kept having issues.... so just passing them as arguments to the menu and then into ur separate functions again too works fine... but it's ugly. 

So you guys are welcome to change things if you find a better way that works... but for now this DOES work until we can ask a TA about a better option or something. So from in my file I can do this now: 

curs.execute("select COUNT(*) from ticket") 
num_tickets = curs.fetchone()[0]
print(num_tickets)

Because it is my understanding that we need to fetch all the info we need via querie or asking the user for input... and then build our insert table statement with that.... correct me if I'm wrong?